make linters and compilers a bit more happy

+++ src/loader/file/meshes.cpp
@@ -16,7 +16,8 @@ RoomStaticMesh RoomStaticMesh::readTr1(io::SDLReader& reader)
   // only in TR2
   room_static_mesh.intensity2 = room_static_mesh.shade.get();
 
-  room_static_mesh.tint.b = room_static_mesh.tint.g = room_static_mesh.tint.r = room_static_mesh.intensity2 / 16384.0f;
+  room_static_mesh.tint.b = room_static_mesh.tint.g = room_static_mesh.tint.r
+    = gsl::narrow_cast<float>(room_static_mesh.intensity2) / 16384.0f;
   room_static_mesh.tint.a = 1.0f;
   return room_static_mesh;
 }
@@ -35,7 +36,8 @@ RoomStaticMesh RoomStaticMesh::readTr2(io::SDLReader& reader)
   if(room_static_mesh.intensity2 >= 0)
     room_static_mesh.intensity2 = (8191 - room_static_mesh.intensity2) * 4;
 
-  room_static_mesh.tint.b = room_static_mesh.tint.g = room_static_mesh.tint.r = room_static_mesh.intensity2 / 16384.0f;
+  room_static_mesh.tint.b = room_static_mesh.tint.g = room_static_mesh.tint.r
+    = gsl::narrow_cast<float>(room_static_mesh.intensity2) / 16384.0f;
   room_static_mesh.tint.a = 1.0f;
   return room_static_mesh;
 }
@@ -48,12 +50,9 @@ RoomStaticMesh RoomStaticMesh::readTr3(io::SDLReader& reader)
   room_static_mesh.shade = core::Shade{reader.readI16()};
   room_static_mesh.intensity2 = reader.readI16();
   room_static_mesh.meshId = core::StaticMeshId::type(reader.readU16());
-
-  room_static_mesh.tint.r = (room_static_mesh.shade.get() & 0x001F) / 62.0f;
-
-  room_static_mesh.tint.g = ((room_static_mesh.shade.get() & 0x03E0) >> 5) / 62.0f;
-
-  room_static_mesh.tint.b = ((room_static_mesh.shade.get() & 0x7C00) >> 10) / 62.0f;
+  room_static_mesh.tint.r = gsl::narrow_cast<float>(room_static_mesh.shade.get() & 0x001F) / 62.0f;
+  room_static_mesh.tint.g = gsl::narrow_cast<float>((room_static_mesh.shade.get() & 0x03E0) >> 5) / 62.0f;
+  room_static_mesh.tint.b = gsl::narrow_cast<float>((room_static_mesh.shade.get() & 0x7C00) >> 10) / 62.0f;
   room_static_mesh.tint.a = 1.0f;
   return room_static_mesh;
 }
@@ -66,12 +65,9 @@ RoomStaticMesh RoomStaticMesh::readTr4(io::SDLReader& reader)
   room_static_mesh.shade = core::Shade{reader.readI16()};
   room_static_mesh.intensity2 = reader.readI16();
   room_static_mesh.meshId = core::StaticMeshId::type(reader.readU16());
-
-  room_static_mesh.tint.r = (room_static_mesh.shade.get() & 0x001F) / 31.0f;
-
-  room_static_mesh.tint.g = ((room_static_mesh.shade.get() & 0x03E0) >> 5) / 31.0f;
-
-  room_static_mesh.tint.b = ((room_static_mesh.shade.get() & 0x7C00) >> 10) / 31.0f;
+  room_static_mesh.tint.r = gsl::narrow_cast<float>(room_static_mesh.shade.get() & 0x001F) / 31.0f;
+  room_static_mesh.tint.g = gsl::narrow_cast<float>((room_static_mesh.shade.get() & 0x03E0) >> 5) / 31.0f;
+  room_static_mesh.tint.b = gsl::narrow_cast<float>((room_static_mesh.shade.get() & 0x7C00) >> 10) / 31.0f;
   room_static_mesh.tint.a = 1.0f;
   return room_static_mesh;
 }
